Remove mixin retina-sprite's $dimensions parameter
At commit 21c5888 a new parameter, $dimensions, was introduced in the central mixin: retina-sprite. Adding $dimensions before the $hover and $active parameters breaks all existing mixin calls which followed the old mixin signature. That change also broke cross-compatibility with @AdamBrodzinski's version, as all clients would need to change all their usages of the retina-sprite mixin if they had already specified arguments for $hover and $active. Moreover, the added parameter was undocumented and caused the demo to fail. If we want to introduce the new parameter, it should be added to the documentation, it's default value should be true, and it should be added as the last parameter to maintain backward compatibility.